[CHORD] Igwe:added PreCertificateToken Test
