+++ package.json
@@ -28,6 +28,8 @@
     "hardhat-gas-reporter": "^1.0.4",
     "hardhat-tracer": "^1.1.1",
     "js-to-ts-converter": "^0.18.1",
+    "keccak256": "^1.0.6",
+    "merkletreejs": "^0.3.3",
     "prettier": "^2.4.1",
     "prettier-plugin-solidity": "^1.0.0-beta.19",
     "solhint": "^3.3.6",
